macos: interface in install
Looks like `fetch` is now `up`. Also, I needed to put an interface on the command line for innernet to work.

This is maybe not exactly the correct fix — at minimum the docs should be updated to say “kick the innernet launchd unit after you set up your interface”. Or perhaps `up` should listen on all interfaces on `/etc/innernet` or something.