adds completions subcommand for the client/server
This subcommand takes a shell as an argument and generates shell
completions for that shell to stdout.

i.e.

```
$ innernet completions bash
  OR
$ innernet-server completions bash
```

This can further be used during packaging to generate shell completions
scripts for the supported shells:

- Bash
- ZSH
- Fish
- PowerShell
- Elvish

Even without packaging, a user could enable completions on the fly:

```
$ source <(innernet completions bash)
```

ZSH requires the completions be in a specific place under the user's
home dir, so the following would allow ZSH users access to
completions (assuming `~/.zfunc` is in their `$FPATH` or substituting
for an appropriate dir if not):

```
$ innernet completions zsh >
~/.zfunc/_innernet && exec zsh
```